backend/requirements.txt
# file to store all the dependencies, pip install -r requirements.txt in correct environment
asgiref
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv